Extract PublishArtifactLocalArtifactMetadataSerializer for reuse
